feat(tui): add Japanese localization for the help overlay
Adds rust-i18n-backed Japanese translations for the `?` help
overlay's key-binding descriptions, marker legend, glossary, and
section headings (docs/adr/0055). Every other screen stays
English-only.

- rinkaku-tui/src/locale.rs: a Locale enum and a pure detect_locale
  function implementing POSIX LC_ALL > LC_MESSAGES > LANG precedence,
  unit-tested in isolation from any env read.
- rinkaku-tui/locales/{en,ja}.yml: the translated strings, compiled
  in via rust_i18n::i18n! at the crate root.
- help::HelpContent's const table becomes help_content(locale), since
  rust_i18n::t! allocates and cannot run in const context; Locale is
  threaded explicitly through run_app/TuiSession::run/ui::draw down
  to draw_help_overlay rather than via a global.
- rinkaku/src/main.rs reads LC_ALL/LC_MESSAGES/LANG at the TUI
  composition root and passes the detected Locale in; rinkaku-tui and
  rinkaku-core never read an environment variable themselves.
- help.rs and ui/help_overlay.rs each crossed the file-size watch
  threshold once locale plumbing landed, so their test modules move
  out to help/tests.rs and ui/help_overlay/tests.rs.
- Adds a TestBackend render test proving the Japanese-locale switch
  actually changes rendered output (key binding, marker legend entry,
  glossary entry, section headings), alongside the existing
  English-locale assertions, which pass unmodified.
