+++ rinkaku-tui/locales/en.yml
@@ -1,0 +1,68 @@
+help:
+  overlay_title: "Help (? to close)"
+  section:
+    markers: "Markers"
+    glossary: "Glossary"
+  group:
+    tree_focus: "Tree focus"
+    right_focus: "Right focus"
+    source_view: "Source view"
+    review: "Review"
+    global: "Global"
+  binding:
+    move_cursor: "Move the cursor"
+    expand_collapse_open: "Expand/collapse a directory row, or open a file/symbol row (moves focus right)"
+    expand_collapse_row: "Expand/collapse a directory/file row (never moves focus)"
+    expand_every_row: "Expand every row"
+    collapse_every_row: "Collapse every row"
+    scroll_right_pane_line: "Scroll the right pane by one line"
+    scroll_right_pane_half_page: "Scroll the right pane by half a page"
+    jump_right_pane_top_bottom: "Jump to the top / bottom of the right pane"
+    return_focus_to_tree: "Return focus to the tree"
+    jump_next_hunk: "Jump to the next hunk (Diff pane only)"
+    jump_previous_hunk: "Jump to the previous hunk (Diff pane only)"
+    scroll_source_pane_line: "Scroll the source pane by one line"
+    scroll_source_pane_half_page: "Scroll the source pane by half a page"
+    jump_file_top_bottom: "Jump to the top / bottom of the file"
+    return_to_entry_view: "Return to the entry view"
+    compose_review_note: "Compose a review note over the symbol under the cursor"
+    open_review_notes_list: "Open the review notes list"
+    notes_list_actions: "Notes list: move, export, close, delete"
+    toggle_detail_diff: "Toggle the right pane between Detail and Diff"
+    toggle_blast_radius: "Toggle the right pane to the blast radius of the selected row"
+    toggle_unified_split: "Toggle unified/split (side-by-side) rendering — Diff pane and source view diff overlay alike"
+    toggle_order_mode: "Toggle topological/alphabetical ordering"
+    open_source_view: "Open the source view for the symbol under the cursor"
+    jump_to_callee: "Jump to a callee of the symbol under the cursor"
+    jump_to_caller: "Jump to a caller of the symbol under the cursor"
+    jump_back: "Jump back to the previous location in the jumplist"
+    jump_forward: "Jump forward to the next location in the jumplist"
+    open_pr_in_browser: "Open the current PR's page in a web browser (--pr mode only)"
+    prompt_self_update: "Prompt to self-update (only shown when a newer release is available)"
+    toggle_help_overlay: "Toggle this help overlay"
+    quit: "Quit (esc/q return to the entry view from the source view)"
+  marker:
+    expand_marker: "Expand marker: children shown / hidden (blank = leaf, nothing to expand)"
+    symbol_kind_prefix: "Symbol row's kind prefix, abbreviated from the language's own keyword"
+    added_symbol: "Added symbol"
+    signature_changed_symbol: "Signature-changed symbol"
+    body_only_symbol: "Body-only, unclassified, or test symbol — exists, but carries less review weight"
+    removed_symbol: "Removed symbol"
+    removed_symbol_name: "Removed symbol's name"
+    dependency_cycle: "Directory contains a dependency cycle"
+    risk_marker: "Risk marker: a contract change and a high-fan-in symbol in the same subtree"
+    whole_test_file_badge: "Whole-test-file badge"
+    collapsed_test_group: "Collapsed group of a file's test symbols"
+    skipped_reason: "Reason a file was not analyzed (parse failure, unsupported language, ...)"
+    changed_count: "Changed, non-removed symbols in this subtree"
+    api_count: "Contract changes in this subtree: signature-changed symbols plus removed symbols"
+    fan_in_count: "Sum of used_by counts over every high-fan-in symbol in this subtree"
+    lines_count: "This file's own line count, colored by file-size band (normal/watch/warn/split)"
+    warn_count: "Directory rows: count of Warn-band files in this subtree"
+    split_count: "Directory rows: count of Split-band files in this subtree"
+  glossary:
+    topological_order: "Directories ordered least-depended-on first, foundations last"
+    alphabetical_order: "Directories ordered A-Z, ignoring dependency direction"
+    blast_radius: "The dependency tree rooted at a selected directory or file, showing what would be affected if it changed"
+    cycle: "A dependency loop: two or more symbols depend on each other, so the tree stops and points back to where it first appeared"
+    jumplist: "The history of gd/gr jump locations — ctrl-o/ctrl-i move back/forward through it"
