+++ rinkaku-tui/locales/ja.yml
@@ -1,0 +1,68 @@
+help:
+  overlay_title: "ヘルプ (? で閉じる)"
+  section:
+    markers: "マーカー"
+    glossary: "用語集"
+  group:
+    tree_focus: "ツリーフォーカス"
+    right_focus: "右ペインフォーカス"
+    source_view: "ソースビュー"
+    review: "レビュー"
+    global: "グローバル"
+  binding:
+    move_cursor: "カーソルを移動"
+    expand_collapse_open: "ディレクトリ行を開閉、またはファイル/シンボル行を開く（フォーカスが右に移動）"
+    expand_collapse_row: "ディレクトリ/ファイル行を開閉（フォーカスは移動しない）"
+    expand_every_row: "全行を展開"
+    collapse_every_row: "全行を折りたたむ"
+    scroll_right_pane_line: "右ペインを1行スクロール"
+    scroll_right_pane_half_page: "右ペインを半ページ分スクロール"
+    jump_right_pane_top_bottom: "右ペインの先頭/末尾へジャンプ"
+    return_focus_to_tree: "フォーカスをツリーに戻す"
+    jump_next_hunk: "次の hunk へジャンプ（Diff ペインのみ）"
+    jump_previous_hunk: "前の hunk へジャンプ（Diff ペインのみ）"
+    scroll_source_pane_line: "ソースペインを1行スクロール"
+    scroll_source_pane_half_page: "ソースペインを半ページ分スクロール"
+    jump_file_top_bottom: "ファイルの先頭/末尾へジャンプ"
+    return_to_entry_view: "エントリービューに戻る"
+    compose_review_note: "カーソル位置のシンボルにレビューノートを作成"
+    open_review_notes_list: "レビューノート一覧を開く"
+    notes_list_actions: "ノート一覧: 移動、エクスポート、閉じる、削除"
+    toggle_detail_diff: "右ペインを Detail と Diff で切り替え"
+    toggle_blast_radius: "右ペインを選択行の blast radius 表示に切り替え"
+    toggle_unified_split: "unified/split（左右分割）表示を切り替え — Diff ペインとソースビューの diff オーバーレイ両方に適用"
+    toggle_order_mode: "topological/alphabetical の並び順を切り替え"
+    open_source_view: "カーソル位置のシンボルのソースビューを開く"
+    jump_to_callee: "カーソル位置のシンボルの callee へジャンプ"
+    jump_to_caller: "カーソル位置のシンボルの caller へジャンプ"
+    jump_back: "jumplist の一つ前の位置へ戻る"
+    jump_forward: "jumplist の次の位置へ進む"
+    open_pr_in_browser: "現在の PR のページをブラウザで開く（--pr モードのみ）"
+    prompt_self_update: "self-update を促す（新しいリリースがある場合のみ表示）"
+    toggle_help_overlay: "このヘルプ画面を切り替え"
+    quit: "終了（ソースビューでは esc/q でエントリービューに戻る）"
+  marker:
+    expand_marker: "展開マーカー: 子要素の表示/非表示（空欄はリーフで展開対象なし）"
+    symbol_kind_prefix: "シンボル行の種別プレフィックス（各言語のキーワードを略記）"
+    added_symbol: "追加されたシンボル"
+    signature_changed_symbol: "シグネチャが変更されたシンボル"
+    body_only_symbol: "body のみの変更、未分類、またはテストのシンボル — 存在はするがレビュー上の重要度は低い"
+    removed_symbol: "削除されたシンボル"
+    removed_symbol_name: "削除されたシンボルの名前"
+    dependency_cycle: "ディレクトリ内に依存の循環がある"
+    risk_marker: "リスクマーカー: 同じサブツリー内に契約変更と高 fan-in シンボルが共存"
+    whole_test_file_badge: "ファイル全体がテストであることを示すバッジ"
+    collapsed_test_group: "ファイル内のテストシンボルをまとめた折りたたみグループ"
+    skipped_reason: "ファイルが解析されなかった理由（パース失敗、未対応言語など）"
+    changed_count: "このサブツリー内の変更済み（削除以外）シンボル数"
+    api_count: "このサブツリー内の契約変更: シグネチャ変更シンボルと削除シンボルの合計"
+    fan_in_count: "このサブツリー内の高 fan-in シンボルの used_by 数の合計"
+    lines_count: "このファイル自体の行数（file-size band: normal/watch/warn/split で色分け）"
+    warn_count: "ディレクトリ行: このサブツリー内の Warn band ファイル数"
+    split_count: "ディレクトリ行: このサブツリー内の Split band ファイル数"
+  glossary:
+    topological_order: "依存が少ないディレクトリから並べる順序（基盤となるものほど後ろ）"
+    alphabetical_order: "依存関係を無視して A-Z で並べる順序"
+    blast_radius: "選択したディレクトリ/ファイルを起点とする依存ツリー — 変更した場合に影響する範囲を示す"
+    cycle: "依存の循環: 複数のシンボルが互いに依存し合っており、ツリーが最初に現れた地点を指して止まる"
+    jumplist: "gd/gr のジャンプ履歴 — ctrl-o/ctrl-i で前後に移動できる"
