+++ rinkaku-tui/src/ui/detail_pane_tests/scroll_behavior.rs
@@ -26,6 +26,7 @@ fn should_show_overflow_indicator_in_detail_pane_title_when_content_exceeds_view
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -59,6 +60,7 @@ fn should_not_show_overflow_indicator_when_content_fits_the_viewport() {
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -94,6 +96,7 @@ fn should_scroll_detail_pane_content_down_when_scroll_down_is_pressed() {
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -135,6 +138,7 @@ fn should_clamp_detail_pane_scroll_at_the_last_page() {
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -177,6 +181,7 @@ fn should_return_the_clamped_scroll_from_draw_when_requested_scroll_overshoots()
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -220,6 +225,7 @@ fn should_reset_scroll_indicator_when_cursor_moves_to_a_different_row() {
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
