+++ rinkaku-tui/src/ui/detail_pane_tests/wrap_reachability.rs
@@ -58,6 +58,7 @@ fn should_reach_the_last_wrapped_line_of_content_via_scrolling_when_a_logical_li
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -111,6 +112,7 @@ fn should_report_indicator_total_as_wrapped_row_count_not_logical_line_count_whe
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
