+++ rinkaku-tui/src/ui/diff_pane_tests/row_kinds.rs
@@ -32,6 +32,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -97,6 +98,7 @@ Binary files a/assets/logo.png and b/assets/logo.png differ
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -168,6 +170,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -237,6 +240,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -301,6 +305,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
