+++ rinkaku-tui/src/ui/diff_pane_tests/styling.rs
@@ -33,6 +33,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -81,6 +82,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -122,6 +124,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -170,6 +173,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
@@ -304,6 +308,7 @@ index e69de29..4b825dc 100644
                 None,
                 &[],
                 &crate::note_markers::NoteMarkers::default(),
+                Locale::English,
             );
         })
         .expect("draw");
