+++ rinkaku-tui/src/ui/overlay.rs
@@ -141,6 +141,7 @@ pub(crate) fn draw_update_prompt(frame: &mut Frame, version: &str, full_area: Re
 #[cfg(test)]
 mod tests {
     use crate::app::{App, BlastRadiusSelection};
+    use crate::locale::Locale;
     use crate::ui::draw;
     use ratatui::Terminal;
     use ratatui::backend::TestBackend;
@@ -229,6 +230,7 @@ mod tests {
                     None,
                     &[],
                     &crate::note_markers::NoteMarkers::default(),
+                    Locale::English,
                 );
             })
             .expect("draw");
@@ -273,6 +275,7 @@ mod tests {
                     None,
                     &[],
                     &crate::note_markers::NoteMarkers::default(),
+                    Locale::English,
                 );
             })
             .expect("draw");
@@ -333,6 +336,7 @@ mod tests {
                         None,
                         &[],
                         &crate::note_markers::NoteMarkers::default(),
+                        Locale::English,
                     );
                 })
                 .expect("draw");
@@ -365,6 +369,7 @@ mod tests {
                     None,
                     &[],
                     &crate::note_markers::NoteMarkers::default(),
+                    Locale::English,
                 );
             })
             .expect("draw");
@@ -393,6 +398,7 @@ mod tests {
                     None,
                     &[],
                     &crate::note_markers::NoteMarkers::default(),
+                    Locale::English,
                 );
             })
             .expect("draw");
@@ -422,6 +428,7 @@ mod tests {
                     None,
                     &[],
                     &crate::note_markers::NoteMarkers::default(),
+                    Locale::English,
                 );
             })
             .expect("draw");
